requirements/base.txt
Django==4.2
environs[django]==9.5.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.5
sorl-thumbnail==12.9.0
Pillow==9.5.0
